+++ dbms/src/Storages/Transaction/TiDB.h
@@ -20,7 +20,8 @@
 #include <IO/WriteHelpers.h>
 #include <Storages/Transaction/StorageEngineType.h>
 #include <Storages/Transaction/Types.h>
-#include <tipb/schema.pb.h>
+#include <TiDB/Schema/DBInfo.h>
+#include <TiDB/Schema/TableInfo.h>
 
 #include <optional>
 
@@ -30,7 +31,6 @@
 #include <Poco/JSON/Array.h>
 #include <Poco/JSON/Object.h>
 #include <Poco/JSON/Parser.h>
-#include <tipb/expression.pb.h>
 #pragma GCC diagnostic pop
 
 namespace DB::ErrorCodes
@@ -270,23 +270,6 @@ struct PartitionInfo
     UInt64 num = 0;
 };
 
-struct DBInfo
-{
-    DatabaseID id = -1;
-    String name;
-    String charset;
-    String collate;
-    SchemaState state;
-
-    DBInfo() = default;
-    explicit DBInfo(const String & json) { deserialize(json); }
-
-    String serialize() const;
-
-    void deserialize(const String & json_str);
-};
-
-struct TableInfo;
 using TableInfoPtr = std::shared_ptr<TableInfo>;
 
 struct TiFlashReplicaInfo
@@ -343,77 +326,10 @@ struct IndexInfo
     bool is_global;
 };
 
-struct TableInfo
+enum class TiFlashMode
 {
-    TableInfo() = default;
-
-    TableInfo(const TableInfo &) = default;
-
-    TableInfo & operator=(const TableInfo &) = default;
-
-    explicit TableInfo(Poco::JSON::Object::Ptr json);
-
-    explicit TableInfo(const String & table_info_json);
-
-    String serialize() const;
-
-    void deserialize(const String & json_str);
-
-    void deserialize(Poco::JSON::Object::Ptr obj);
-
-    // The meaning of this ID changed after we support TiDB partition table.
-    // It is the physical table ID, i.e. table ID for non-partition table,
-    // and partition ID for partition table,
-    // whereas field `belonging_table_id` below actually means the table ID this partition belongs to.
-    TableID id = DB::InvalidTableID;
-    String name;
-    // Columns are listed in the order in which they appear in the schema.
-    std::vector<ColumnInfo> columns;
-    /// index_infos stores the index info from TiDB. But we do not store all
-    /// the index infos because most of the index info is useless in TiFlash.
-    /// If is_common_handle = true, the primary index info is stored
-    /// otherwise, all of the index info are ignored
-    std::vector<IndexInfo> index_infos;
-    SchemaState state = StateNone;
-    bool pk_is_handle = false;
-    /// when is_common_handle = true, it means this table is a clustered index table
-    bool is_common_handle = false;
-    String comment;
-    Timestamp update_timestamp = 0;
-    bool is_partition_table = false;
-    TableID belonging_table_id = DB::InvalidTableID;
-    PartitionInfo partition;
-    // If the table is view, we should ignore it.
-    bool is_view = false;
-    // If the table is sequence, we should ignore it.
-    bool is_sequence = false;
-    Int64 schema_version = DEFAULT_UNSPECIFIED_SCHEMA_VERSION;
-
-    // The TiFlash replica info persisted by TiDB
-    TiFlashReplicaInfo replica_info;
-
-    ::TiDB::StorageEngine engine_type = ::TiDB::StorageEngine::UNSPECIFIED;
-
-    ColumnID getColumnID(const String & name) const;
-    String getColumnName(ColumnID id) const;
-
-    const ColumnInfo & getColumnInfo(ColumnID id) const;
-
-    std::optional<std::reference_wrapper<const ColumnInfo>> getPKHandleColumn() const;
-
-    TableInfoPtr producePartitionTableInfo(TableID table_or_partition_id, const DB::SchemaNameMapper & name_mapper) const;
-
-    bool isLogicalPartitionTable() const { return is_partition_table && belonging_table_id == DB::InvalidTableID && partition.enable; }
-
-    /// should not be called if is_common_handle = false.
-    /// when use IndexInfo, please avoid to use the offset info
-    /// the offset value may be wrong in some cases,
-    /// due to we will not update IndexInfo except RENAME DDL action,
-    /// but DDL like add column / drop column may change the offset of columns
-    /// Thus, please be very careful when you must have to use offset information !!!!!
-    const IndexInfo & getPrimaryIndexInfo() const { return index_infos[0]; }
-
-    IndexInfo & getPrimaryIndexInfo() { return index_infos[0]; }
+    Normal,
+    Fast,
 };
 
 using DBInfoPtr = std::shared_ptr<DBInfo>;
