+++ dbms/src/TiDB/Schema/DBInfo.cpp
@@ -1,0 +1,67 @@
+// Copyright 2022 PingCAP, Ltd.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+#include <Common/Exception.h>
+#include <Poco/Types.h>
+#include <TiDB/Schema/DBInfo.h>
+#include <common/types.h>
+
+namespace TiDB
+{
+String DBInfo::serialize() const
+try
+{
+    std::stringstream buf;
+
+    Poco::JSON::Object::Ptr json = new Poco::JSON::Object();
+    json->set("id", id);
+    Poco::JSON::Object::Ptr name_json = new Poco::JSON::Object();
+    name_json->set("O", name);
+    name_json->set("L", name);
+    json->set("db_name", name_json);
+
+    json->set("charset", charset);
+    json->set("collate", collate);
+
+    json->set("state", static_cast<Int32>(state));
+
+    json->stringify(buf);
+
+    return buf.str();
+}
+catch (const Poco::Exception & e)
+{
+    throw DB::Exception(
+        std::string(__PRETTY_FUNCTION__) + ": Serialize TiDB schema JSON failed (DBInfo): " + e.displayText(),
+        DB::Exception(e));
+}
+
+void DBInfo::deserialize(const String & json_str)
+try
+{
+    Poco::JSON::Parser parser;
+    Poco::Dynamic::Var result = parser.parse(json_str);
+    auto obj = result.extract<Poco::JSON::Object::Ptr>();
+    id = obj->getValue<DatabaseID>("id");
+    name = obj->get("db_name").extract<Poco::JSON::Object::Ptr>()->get("L").convert<String>();
+    charset = obj->get("charset").convert<String>();
+    collate = obj->get("collate").convert<String>();
+    state = static_cast<SchemaState>(obj->getValue<Int32>("state"));
+}
+catch (const Poco::Exception & e)
+{
+    throw DB::Exception(
+        std::string(__PRETTY_FUNCTION__) + ": Parse TiDB schema JSON failed (DBInfo): " + e.displayText() + ", json: " + json_str,
+        DB::Exception(e));
+}
+} // namespace TiDB
