+++ dbms/src/TiDB/Schema/DBInfo.h
@@ -1,0 +1,36 @@
+// Copyright 2022 PingCAP, Ltd.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include <Storages/Transaction/TiDB.h>
+
+namespace TiDB
+{
+struct DBInfo
+{
+    DatabaseID id = -1;
+    String name;
+    String charset;
+    String collate;
+    SchemaState state;
+
+    DBInfo() = default;
+    explicit DBInfo(const String & json) { deserialize(json); }
+
+    String serialize() const;
+
+    void deserialize(const String & json_str);
+};
+} // namespace TiDB
